(WIP) Enable update on distribution column for legacy planner.
For legacy planner, it will reject updating on distribution columns. In our fix,
we leverage SplitUpdate node which is used by ORCA, and decorate plan by adding SplitUpdate node
in phase of ModifyTable. In SplitUpdate, it use old attributes as routing to segment for deletion.
So we need to get both new attributes and old attributes. If any old attributes absent, we should
add it to targetlist recursivly.

TODO:
* Fix several cases change caused on enabling update on distribution columns. We don't need an extra
  distribution column now.
* Fix width in plan node to reflect really width since we add some extra old attributes.

Author: Xiaoran Wang <xiwang@pivotal.io>
Author: Max Yang <myang@pivotal.io>